silence some warnings

+++ bam_import.c
@@ -177,14 +177,15 @@ static inline void append_text(bam_header_t *header, kstring_t *str)
         header->text = (char*)realloc(header->text, y);
         if ( !header->text ) 
         {
-            fprintf(stderr,"realloc failed to alloc %ld bytes\n", y);
+            fprintf(stderr,"realloc failed to alloc %ld bytes\n", (long)y);
             abort();
         }
     }
     // Sanity check
     if ( header->l_text+str->l+1 >= header->n_text )
     {
-        fprintf(stderr,"append_text FIXME: %ld>=%ld, x=%ld,y=%ld\n",  header->l_text+str->l+1,header->n_text,x,y);
+        fprintf(stderr,"append_text FIXME: %ld>=%ld, x=%ld,y=%ld\n",
+                (long)header->l_text+str->l+1,(long)header->n_text,(long)x,(long)y);
         abort();
     }
 	strncpy(header->text + header->l_text, str->s, str->l+1); // we cannot use strcpy() here.
